backend/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.118.0
furl==2.1.4
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
orderedmultidict==1.0.1
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-mock==3.15.1
python-dotenv==1.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
